pyproject.toml
[project]
name = "project"
version = "0.1.0"
requires-python = ">=3.12, <3.13"
dependencies = [
  "torch",
  "torchvision",
  "matplotlib",
  "docling",
  "transformers",
  "sentence-transformers>=5.4.0",
  "numpy<2",
  "mlflow>=3.11.1",
  "pytest>=9.0.3",
  "faiss-cpu>=1.13.2",
  "gradio>=6.14.0",
  "coverage>=7.13.5",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true